llamaindex/pyproject.toml
[tool.poetry]
name = "llamaindex"
version = "0.1.0"
description = ""
authors = ["Pierre"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.2"
llama-index = {version = "0.9.7", python = ">=3.11,<3.12"}
python-dotenv = "1.0.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
